chore(@gui files): split settings screen file to separate settings screens (as in UI)
