+++ gui/components/profile_popup.py
@@ -7,7 +7,7 @@ from gui.components.base_popup import BasePopup
 from gui.elements.button import Button
 from gui.elements.object import QObject
 from gui.elements.text_label import TextLabel
-from gui.screens.settings import ProfileSettingsView
+from gui.screens.settings_profile import ProfileSettingsView
 from scripts.tools.image import Image
